refactor(core): move session-title into internal, narrow its public surface
git-mv src/session-title.ts to src/internal/session-title.ts (alongside
dates.ts / session-support.ts), fix its relative imports, and mark the header
as internal. session.ts now imports generateTitleWithLLM from the internal
path.

Public surface narrowed at the barrel: buildTitlePrompt and
generateTitleWithLLM are no longer exported (core-internal assembly details
of Session.generateTitle; the core test imports them by relative path).
SessionTitleResult (part of Session.generateTitle's public signature),
stripConversationMarkers AND sanitizeTitle stay public as a documented
narrow re-export from the internal module — the server's title fallback
(runtime/title-generator.ts) uses BOTH sanitation helpers, so dropping
sanitizeTitle as originally planned would have broken it; keeping it public
lets the server compile completely unchanged. The re-export layout (rather
than relocating declarations into session.ts) avoids a runtime import cycle
between session.ts and the internal module and keeps the declarations next
to the logic they describe.

No behavior change; the tsup entry list is untouched (internal files bundle
transitively through the barrel).
